update 1.1
activation added
views classes moved to views folder
