--- core/accounts/api/v1/views.py
@@ -1,116 +1,0 @@
-from rest_framework import generics
-from rest_framework.response import Response
-from rest_framework import status
-from .serializers import (
-                RegistrationSerializer,
-                CustomAuthTokenSerializer,
-                CustomTokenObtainPairSerializer,
-                ChangePasswordApiSerializer,
-                ProfileSerializer,)
-from rest_framework.authtoken.views import ObtainAuthToken
-from rest_framework.authtoken.models import Token
-from rest_framework.views import APIView
-from rest_framework.permissions import IsAuthenticated
-# jwt
-from rest_framework_simplejwt.views import TokenObtainPairView
-# change password
-from django.contrib.auth import get_user_model
-# profile
-from ...models import Profile
-from django.shortcuts import get_object_or_404
-# email
-# from django.core.mail import send_mail
-# from mail_templated import send_mail
-from mail_templated import EmailMessage
-from ..utils import EmailThread
-
-User = get_user_model()
-
-
-
-class RegistrationApiView(generics.GenericAPIView):
-    serializer_class = RegistrationSerializer
-
-    def post(self, request, *args, **kwargs):
-        serializer = RegistrationSerializer(data= request.data)
-        if serializer.is_valid():
-            serializer.save()
-            data ={
-                'detail' : 'completed registration',
-                'email': serializer.validated_data['email'],
-                
-            }
-            return Response(data, status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-
-# custom Auth Token
-class CustomAuthToken(ObtainAuthToken):
-    serializer_class = CustomAuthTokenSerializer
-    def post(self, request, *args, **kwargs):
-        serializer = self.serializer_class(data=request.data,
-                                           context={'request':request})
-        serializer.is_valid(raise_exception=True)
-        user = serializer.validated_data['user']
-        token, created = Token.objects.get_or_create(user=user)
-        return Response({
-            'token': token.key,
-            'user_id': user.pk,
-            'email': user.email
-        })
-
-
-# discard token
-class CustomDiscardAuthToken(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request, *args, **kwargs):
-        request.user.auth_token.delete()
-        return Response(status= status.HTTP_204_NO_CONTENT)
-    
-
-# custom TokenObtainPairView
-class CostumTokenObtainPairView(TokenObtainPairView):
-    serializer_class = CustomTokenObtainPairSerializer
-
-
-# change password view
-class ChangePasswordApiView(generics.GenericAPIView):
-    model = User
-    permission_classes = [IsAuthenticated,]
-    serializer_class = ChangePasswordApiSerializer
-
-    def get_object(self, queryset=None):
-        obj = self.request.user
-        return obj
-    def put(self, request, *args, **kwargs):
-        self.object = self.get_object()
-        serializer = self.get_serializer(data=request.data)
-        if serializer.is_valid():
-            # check old password
-            if not self.object.check_password(serializer.data.get('old_password')):
-                return Response({"old Password": "Wrong password"}, status=status.HTTP_400_BAD_REQUEST)
-            # set_password also hashes the password that the user wiil get
-            self.object.set_password(serializer.data.get('new_password'))
-            self.object.save()
-            return Response({'details':'Password changed successfully'}, status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-# profile
-class ProfileApiView(generics.RetrieveUpdateAPIView):
-    serializer_class = ProfileSerializer
-    queryset = Profile.objects.all()
-
-    def get_object(self):
-        queryset = self.get_queryset()
-        obj = get_object_or_404(queryset, user=self.request.user)
-        return obj
-    
-
-# email
-class EmailTestSend(generics.GenericAPIView):
-    def post(self, *args, **kwargs):
-        profile = Profile.objects.get(user=self.request.user.id)
-        email_obj = EmailMessage('email/hello.tpl', {'name': profile.first_name }, 'admin@admin.com', to=[profile.user.email] )
-        EmailThread(email_obj)
-        return Response('email sent')
