+++ core/accounts/api/v1/views/CustomDiscardAuthToken.py
@@ -1,0 +1,17 @@
+from rest_framework.response import Response
+from rest_framework import status
+from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
+
+from django.contrib.auth import get_user_model
+User = get_user_model()
+
+
+# discard token
+class CustomDiscardAuthToken(APIView):
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request, *args, **kwargs):
+        request.user.auth_token.delete()
+        return Response(status= status.HTTP_204_NO_CONTENT)
+
