+++ core/accounts/api/v1/views/RegistrationApiView.py
@@ -1,0 +1,47 @@
+from rest_framework import generics
+from rest_framework.response import Response
+from rest_framework import status
+from ..serializers import RegistrationSerializer
+from django.shortcuts import get_object_or_404
+# email
+from mail_templated import EmailMessage
+from ..utils import EmailThread
+from rest_framework_simplejwt.tokens import RefreshToken
+
+
+from django.contrib.auth import get_user_model
+User = get_user_model()
+
+
+
+class RegistrationApiView(generics.GenericAPIView):
+    serializer_class = RegistrationSerializer
+
+    def post(self, request, *args, **kwargs):
+        serializer = RegistrationSerializer(data= request.data)
+        if serializer.is_valid():
+            serializer.save()
+            email = serializer.validated_data['email']
+            data ={
+                'detail' : 'We sent an email to you for verification. Please check your email and click the link to verify your account.',
+                'email': email,
+                
+            }
+            user_obj = get_object_or_404(User, email=email)
+            token = self.get_tokens_for_user(user_obj)
+            email_obj = EmailMessage(
+                'email/activation_email.tpl',
+                {'token': token },
+                'admin@admin.com', to=[email]
+            )
+            # multi threading
+            EmailThread(email_obj).start()
+            return Response(data, status.HTTP_201_CREATED)
+        
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+    # get access token for user
+    def get_tokens_for_user(self, user):
+        refresh = RefreshToken.for_user(user)
+        return str(refresh.access_token)
+
