+++ core/accounts/api/v1/views/__init__.py
@@ -1,0 +1,9 @@
+# api-v1 views
+from .RegistrationApiView import *
+from .ActivationApiView import *
+from .ChangePasswordApiView import *
+from .CostumTokenObtainPairView import *
+from .CustomAuthToken import *
+from .CustomDiscardAuthToken import *
+from .EmailTestSend import *
+from .ProfileApiView import *
