+++ core/templates/email/activation_email.tpl
@@ -1,0 +1,9 @@
+{% extends "mail_templated/base.tpl" %}
+
+{% block subject %}
+Account Activation
+{% endblock %}
+
+{% block html %}
+<a href="http://127.0.0.1:8000/accounts/api/v1/activation/confirm/{{token}}">Verify your user</a>
+{% endblock %}
